terry/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
boto==2.48.0
botocore==1.7.35
celery==4.0.2
Django==1.10.5
django-allauth==0.30.0
django-paypal==0.3.6
django-watchman==0.13.1
djangorestframework==3.5.3
dj-database-url==0.4.1
enum34==1.1.6
enum-compat==0.0.2
eventlet==0.20.1
GitPython==2.1.1
greenlet==0.4.12
gunicorn==19.6.0
itsdangerous==0.24
jsonfield==2.0.1
kombu==4.0.2
psycopg2==2.7.3
pycurl==7.43.0
pygooglechart==0.4.0
PyOpenSSL==17.3.0
python-dotenv==0.7.1
python-hglib==2.3
python-http-client==2.2.1
pytz==2016.10
redis==2.10.5
sendgrid==3.6.3
sendgrid-django==4.0.4
vine==1.1.3
WeasyPrint==0.36

-r snoopy/requirements.txt
